Fix for install.bash to create /etc/ros directory if necessary.

+++ husky_bringup/upstart/install.bash
@@ -13,7 +13,14 @@ export release=$(ls /opt/ros/ | tail -n1)
 
 source helpers.bash
 
-rm -f /etc/ros/setup.bash
+if [ -d /etc/ros ]; then
+  # Remove existing setup file so we can use our own.
+  rm -f /etc/ros/setup.bash
+else
+  # Create /etc/ros so that we can put our file in there.
+  mkdir /etc/ros
+fi
+
 ln -s /home/${user}/ros/setup.bash /etc/ros/setup.bash
 source /etc/ros/setup.bash
 pushd `rospack find ${robot}_bringup`/upstart > /dev/null
